Maturity not showing on My Neuron webpage
As you can comprove, my neuron have 194 ICP + 12 of Already merged maturity (automatically) = 206 icp for rewards calculation, and 0.11 icp per day. And until yesterday everything was OK, but today, the merged staked maturity is not showing anymore at the website.

The only amount that is showing is 194 icp

https://dashboard.internetcomputer.org/neuron/1151775483429906163

So the point is... What the hell is going on?